0x0A

+++ 0x0A-argc_argv/3-mul.c
@@ -11,8 +11,8 @@ if (argc == 3)
 int number_1;
 int number_2;
 int product;
-number_1 = (int)*argv[1];
-number_2 = (int)*argv[2];
+number_1 = *argv[1];
+number_2 = *argv[2];
 product = ((number_1 - '0') * (number_2 - '0'));
 printf("%d\n", product);
 return (0);
